requirements.txt
fastapi
uvicorn
razorpay
python-dotenv
anyio<4.0.0
